+++ 1.21.3/Fabric/src/main/java/fuzs/puzzleslib/fabric/impl/core/FabricClientProxy.java
@@ -7,6 +7,7 @@ import fuzs.puzzleslib.impl.network.codec.CustomPacketPayloadAdapter;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
+import net.minecraft.client.multiplayer.MultiPlayerGameMode;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.protocol.Packet;
@@ -26,7 +27,7 @@ public class FabricClientProxy extends FabricServerProxy implements ClientProxyI
     public <M1, M2> void registerClientReceiver(CustomPacketPayload.Type<CustomPacketPayloadAdapter<M1>> type, BiConsumer<Throwable, Consumer<Component>> disconnectExceptionally, Function<M1, ClientboundMessage<M2>> messageAdapter) {
         ClientPlayNetworking.registerGlobalReceiver(type,
                 (CustomPacketPayloadAdapter<M1> payload, ClientPlayNetworking.Context context) -> {
-                    context.client().submit(() -> {
+                    try {
                         Objects.requireNonNull(context.player(), "player is null");
                         ClientboundMessage<M2> message = messageAdapter.apply(payload.unwrap());
                         message.getHandler()
@@ -34,26 +35,26 @@ public class FabricClientProxy extends FabricServerProxy implements ClientProxyI
                                         context.client(),
                                         context.player().connection,
                                         context.player(),
-                                        context.client().level
-                                );
-                    }).exceptionally((Throwable throwable) -> {
+                                        context.client().level);
+                    } catch (Throwable throwable) {
                         disconnectExceptionally.accept(throwable, context.responseSender()::disconnect);
-                        return null;
-                    });
-                }
-        );
+                    }
+                });
     }
 
     @Override
     public boolean shouldStartDestroyBlock(BlockPos blockPos) {
-        MultiPlayerGameModeFabricAccessor gameMode = (MultiPlayerGameModeFabricAccessor) Minecraft.getInstance().gameMode;
-        return !gameMode.puzzleslib$getIsDestroying() || !gameMode.puzzleslib$callSameDestroyTarget(blockPos);
+        MultiPlayerGameMode gameMode = Minecraft.getInstance().gameMode;
+        Objects.requireNonNull(gameMode, "game mode is null");
+        return !gameMode.isDestroying() ||
+                !((MultiPlayerGameModeFabricAccessor) gameMode).puzzleslib$callSameDestroyTarget(blockPos);
     }
 
     @Override
     public void startClientPrediction(Level level, IntFunction<Packet<ServerGamePacketListener>> predictiveAction) {
-        ((MultiPlayerGameModeFabricAccessor) Minecraft.getInstance().gameMode).puzzleslib$callStartPrediction((ClientLevel) level,
-                predictiveAction::apply
-        );
+        MultiPlayerGameMode gameMode = Minecraft.getInstance().gameMode;
+        Objects.requireNonNull(gameMode, "game mode is null");
+        ((MultiPlayerGameModeFabricAccessor) gameMode).puzzleslib$callStartPrediction((ClientLevel) level,
+                predictiveAction::apply);
     }
 }